Fixed a bug with background elements.
Cucumber.js supports background steps that get prepended to every
scenario. The set of background steps are also included as an element in
the JSON output, but the steps in it aren't executed there, so they
don't have a result property and thus make cucumber-junit crash.

This filters out elements of type background. That's ok. Their result of
their steps is added to each scenario where those steps are executed.